+++ telegrambots-meta/src/main/java/org/telegram/telegrambots/api/methods/send/SendMessage.java
@@ -14,11 +14,12 @@ import org.telegram.telegrambots.exceptions.TelegramApiValidationException;
 import java.io.IOException;
 import java.util.Objects;
 
+import static com.google.common.base.Preconditions.checkNotNull;
+
 /**
  * @author Ruben Bermudez
  * @version 1.0
- * @brief Use this method to send text messages. On success, the sent Message is returned.
- * @date 20 of June of 2015
+ * Use this method to send text messages. On success, the sent Message is returned.
  */
 public class SendMessage extends BotApiMethod<Message> {
     public static final String PATH = "sendmessage";
@@ -48,6 +49,16 @@ public class SendMessage extends BotApiMethod<Message> {
 
     public SendMessage() {
         super();
+    }
+
+    public SendMessage(String chatId, String text) {
+        this.chatId = checkNotNull(chatId);
+        this.text = checkNotNull(text);
+    }
+
+    public SendMessage(Integer chatId, String text) {
+        this.chatId = checkNotNull(chatId).toString();
+        this.text = checkNotNull(text);
     }
 
     public String getChatId() {